Post-login steps in base fragment, prior to epilogue
