+++ WordPress/src/main/java/org/wordpress/android/ui/accounts/login/LoginEmailPasswordFragment.java
@@ -197,9 +197,9 @@ public class LoginEmailPasswordFragment extends LoginBaseFormFragment implements
     @SuppressWarnings("unused")
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onAuthenticationChanged(OnAuthenticationChanged event) {
-        endProgress();
-
         if (event.isError()) {
+            endProgress();
+
             AppLog.e(T.API, "onAuthenticationChanged has error: " + event.error.type + " - " + event.error.message);
 
             if (isAdded()) {
@@ -211,9 +211,11 @@ public class LoginEmailPasswordFragment extends LoginBaseFormFragment implements
 
         AppLog.i(T.NUX, "onAuthenticationChanged: " + event.toString());
 
-        if (mLoginListener != null) {
-            mLoginListener.loggedInViaPassword(mOldSitesIDs);
-        }
+        doFinishLogin();
     }
 
+    @Override
+    protected void onLoginFinished() {
+        mLoginListener.loggedInViaPassword(mOldSitesIDs);
+    }
 }
